// src/enums/PolarizationType.ts
/**
 * @author @thkruz/ootk
 * @license AGPL-3.0-or-later
 * @copyright (c) 2024 Theodore Kruczek
 *
 * Orbital Object ToolKit is free software: you can redistribute it and/or modify it under the
 * terms of the GNU Affero General Public License as published by the Free Software
 * Foundation, either version 3 of the License, or (at your option) any later version.
 *
 * Orbital Object ToolKit is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
 * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
 * See the GNU Affero General Public License for more details.
 *
 * You should have received a copy of the GNU Affero General Public License along with
 * Orbital Object ToolKit. If not, see <http://www.gnu.org/licenses/>.
 */

/**
 * Antenna polarization types.
 */
export enum PolarizationType {
  /** Linear horizontal polarization */
  LINEAR_HORIZONTAL = 'LINEAR_HORIZONTAL',
  /** Linear vertical polarization */
  LINEAR_VERTICAL = 'LINEAR_VERTICAL',
  /** Right-hand circular polarization */
  RHCP = 'RHCP',
  /** Left-hand circular polarization */
  LHCP = 'LHCP',
  /** Dual linear polarization */
  DUAL_LINEAR = 'DUAL_LINEAR',
  /** Dual circular polarization */
  DUAL_CIRCULAR = 'DUAL_CIRCULAR',
  /** Elliptical polarization */
  ELLIPTICAL = 'ELLIPTICAL',
}